# doctorbooking/src/main/resources/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: My API
  description: This is a sample Spring Boot RESTful service using springdoc-openapi
    and OpenAPI 3.
  version: "1.0"
servers:
- url: http://localhost:8089
  description: Generated server url
paths:
...
  /api/doctors/slots:
    get:
      tags:
      - slots-availability-controller
      operationId: getAllSlots
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SlotDTO"
    post:
      tags:
      - slots-availability-controller
      operationId: createSlot
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SlotDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    SlotDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
        doctorName:
          type: string
        isReserved:
          type: boolean
        dateTime:
          type: string
        cost:
          type: number
          format: double